requirements.txt
# requirements.txt
# Core
pyyaml==6.0.1
networkx==3.2.1
ollama==0.1.6

# Парсинг
python-hcl2==4.3.2
javalang==0.13.0

# Utilities
gitpython==3.1.40
pathlib==1.0.1
dataclasses-json==0.6.3

# Optional
plantuml==0.3.0
#pygraphviz==1.11